+++ protos/vega/vega.pb.go
@@ -8284,7 +8284,7 @@ type ReferralProgram struct {
 	// second element will give Tier 2, and so on. Determines the level of
 	// benefit a party can expect based on performance criteria.
 	BenefitTiers []*BenefitTier `protobuf:"bytes,3,rep,name=benefit_tiers,json=benefitTiers,proto3" json:"benefit_tiers,omitempty"`
-	// Timestamp as Unix time in seconds after which when the current epoch ends, the
+	// Timestamp as Unix time in seconds, after which when the current epoch ends, the
 	// programs status will become STATE_CLOSED and benefits will be disabled.
 	EndOfProgramTimestamp int64 `protobuf:"varint,4,opt,name=end_of_program_timestamp,json=endOfProgramTimestamp,proto3" json:"end_of_program_timestamp,omitempty"`
 	// Number of epochs over which to evaluate a referral set's running volume.